Update API requirements -> requirement and updated return value of min_nqf_level in API
