+++ django_project/api/serializers/requirements.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
 from ford3.enums.saqa_qualification_level import SaqaQualificationLevel
 from ford3.models.requirement import Requirement
-from api.serializers.utilities.fields import EnumField
+from api.serializers.utilities.enum_field import EnumField
 
 
 class RequirementSerializer(serializers.ModelSerializer):
